chore: make mimetype detection more reliable
This commit introduces a new function, `getFullMimeType`, which uses the `mimetype` package to detect the MIME type of a file. This function is used to set the `FullMimeType` field of the `db.Video` and `db.Image` structs, ensuring that the MIME type is always accurate.

The previous method of setting the `FullMimeType` field relied on the `file.MimeType` field, which was not always reliable. This could lead to incorrect MIME types being stored in the database.

The new `getFullMimeType` function is more reliable because it uses the `mimetype` package, which is a well-maintained and widely used library for MIME type detection.