feat: add mime_type filter to files route
Adds a new query parameter, `type`, to the `GET /files` route. This parameter allows the user to filter the results by the `mime_type` of the file.

This feature is useful for applications that need to display files of a specific type, such as images or videos. It allows the user to easily filter the results without having to manually search through all of the files.