feat: Add caching to optimize performance
This change introduces a cache layer to improve the performance of the application by reducing the number of database queries. The cache is used to store frequently accessed data, such as the list of files and the details of a single file, which can significantly reduce the response time for these requests.